Saving Data to Local Storage
I made a function to save the city name searched in form search bar to local storage along. I also saved the search history to the webpage so you can access later. So the city data and the search hsitory is saved to local storage meaning the openweather api data for the city is saved along with the searched item in form of a button.