Annotation
I added annotation to the html and css. I already added the annotation to the javascript as I completed each section and confirmed that each section worked. I am going to continue doing this as its so much easier to follow along what I'm doing if I already have notes.